statsd: collapse the clean helpers into statsd helper
