+++ lib/statsd.js
@@ -24,7 +24,11 @@ var assert = require('assert');
 
 var STAT_EMIT_PERIOD = 100;
 
+BatchStatsd.clean = clean;
+BatchStatsd.cleanHostPort = cleanHostPort;
 BatchStatsd.BaseStat = BaseStat;
+
+var LENGTH_ARRAYS = {};
 
 function BatchStatsd(options) {
     if (!(this instanceof BatchStatsd)) {
@@ -151,4 +155,64 @@ function StatTags(opts) {
             self.version = opts.version;
         }
     }
+}
+
+function clean(str, field) {
+    var copy;
+
+    if (!str) {
+        return field;
+    }
+
+    copy = LENGTH_ARRAYS[str.length];
+    if (!copy) {
+        copy = LENGTH_ARRAYS[str.length] = [];
+    }
+
+    for (var i = 0; i < str.length; i++) {
+        var char = str[i];
+
+        if (char === ':' ||
+            char === '/' ||
+            char === '.' ||
+            char === '{' ||
+            char === '}'
+        ) {
+            copy[i] = '-';
+        } else {
+            copy[i] = char;
+        }
+    }
+
+    return copy.join('');
+}
+
+function cleanHostPort(str, field) {
+    var copy;
+
+    if (!str) {
+        return field;
+    }
+
+    var length = str.indexOf(':');
+
+    copy = LENGTH_ARRAYS[length];
+    if (!copy) {
+        copy = LENGTH_ARRAYS[length] = [];
+    }
+    for (var i = 0; i < length; i++) {
+        var char = str[i];
+
+        if (char === '/' ||
+            char === '.' ||
+            char === '{' ||
+            char === '}'
+        ) {
+            copy[i] = '-';
+        } else {
+            copy[i] = char;
+        }
+    }
+
+    return copy.join('');
 }
